requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
beautifulsoup4==4.7.1
cffi==1.12.3
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
Django==2.2.1
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.9.4
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.3
model-mommy==1.6.0
mongoengine==0.17.0
parsel==1.5.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.8.0
pyOpenSSL==19.0.0
pypiwin32==223
pytz==2019.1
pywin32==224
queuelib==1.5.0
Scrapy==1.6.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
Twisted==19.2.0
w3lib==1.20.0
zope.interface==4.6.0
